refactor: move naming generation methods back to ControllerUtils
