+++ operator-framework-quarkus-extension/deployment/src/main/java/io/javaoperatorsdk/quarkus/extension/deployment/QuarkusExtensionProcessor.java
@@ -144,8 +144,7 @@ class QuarkusExtensionProcessor {
                 "name",
                 AnnotationValue::asString,
                 () ->
-                    ResourceController.getDefaultResourceControllerName(
-                        resourceControllerClassName)),
+                    ControllerUtils.getDefaultResourceControllerName(resourceControllerClassName)),
             crdName,
             valueOrDefault(
                 controllerAnnotation,
